Consolidated disparate resume.json files
This now has a single base-resume.json that the others are to be
generated from. For the specified document paths, elements in the lists
are rendered as strings if they are strings, otherwise they are expected
to be dicts where the current generated key's value is extracted if present,
else the "default" value is used if present.